pyproject.toml
[tool.poetry]
name = "nomad-django"
version = "0.1.0"
description = ""
authors = ["haru <godjaewook@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
django = "^5.0.4"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
